Perform migration only if required.
This addresses issues with running both auto-upgrade and explicit
migrations.

@pbadenski
@robyoung

+++ db/migrate/use_recorder_gem_for_daily_visitors.rb
@@ -4,9 +4,17 @@ require "dm-migrations/migration_runner"
 migration 1, :use_recorder_gem_for_daily_visitors do
   up do
     modify_table :daily_unique_visitors do
-      drop_column :updated_at
-      drop_column :created_at
-      add_column :source, String, allow_nil: false
+      if adapter.field_exists?("daily_unique_visitors", "updated_at")
+        drop_column :updated_at
+      end
+
+      if adapter.field_exists?("daily_unique_visitors", "created_at")
+        drop_column :created_at
+      end
+
+      unless adapter.field_exists?("daily_unique_visitors", "source")
+         add_column :source, String, allow_nil: false
+      end 
     end
   end
 end
